+++ src/api/mod.rs
@@ -1,2 +1,3 @@
 pub mod github;
+pub mod huggingface;
 pub mod wakatime;