+++ src/prepared_templates.rs
@@ -5,6 +5,7 @@ use axum::response::{IntoResponse, Response};
 #[derive(Debug)]
 pub enum PreparedTemplate {
     FailedFindUser,
+    FailedFindRepo,
     FailedFindLanguages,
     BadCredentials,
     APIRateLimit,
@@ -16,6 +17,10 @@ impl PreparedTemplate {
         let template = match self {
             PreparedTemplate::FailedFindUser => ErrorTemplate {
                 first_line: "Failed to find a user.",
+                second_line: "Check if it’s spelled correctly",
+            },
+            PreparedTemplate::FailedFindRepo => ErrorTemplate {
+                first_line: "Failed to find a repo.",
                 second_line: "Check if it’s spelled correctly",
             },
             PreparedTemplate::FailedFindLanguages => ErrorTemplate {